Refactor: change package name
Change package name to noun
Change package name to not contain lowercase letters
