+++ internal/job/adapter/factory.go
@@ -1,7 +1,7 @@
-package fetch
+package adapter
 
 import "github.com/Goboolean/core-system.worker/internal/job"
 
-func Create(name string, p *job.UserParams) (Fetcher, error) {
+func Create(name string, p *job.UserParams) (Adapter, error) {
 	panic("not implied")
 }
