+++ internal/job/adapter/interface.go
@@ -1,13 +1,16 @@
-package analyze
+package adapter
 
 import (
 	"github.com/Goboolean/core-system.worker/internal/job"
 )
 
-type Analyzer interface {
+type Adapter interface {
 	job.Common
 
 	SetRefInput(chan any)
 	SetPredictInput(chan any)
 	Output() chan any
 }
+
+// adapt.Common
+// adapt.adaptor
