+++ internal/job/analyzer/factory.go
@@ -1,7 +1,7 @@
-package transmit
+package analyzer
 
 import "github.com/Goboolean/core-system.worker/internal/job"
 
-func Create(name string, p *job.UserParams) (Transmitter, error) {
+func Create(name string, p *job.UserParams) (Analyzer, error) {
 	panic("not implied")
 }
