+++ internal/job/analyzer/interface.go
@@ -1,16 +1,13 @@
-package adapt
+package analyzer
 
 import (
 	"github.com/Goboolean/core-system.worker/internal/job"
 )
 
-type Adapter interface {
+type Analyzer interface {
 	job.Common
 
 	SetRefInput(chan any)
 	SetPredictInput(chan any)
 	Output() chan any
 }
-
-// adapt.Common
-// adapt.adaptor
