+++ internal/job/executer/factory.go
@@ -1,7 +1,7 @@
-package adapt
+package executer
 
 import "github.com/Goboolean/core-system.worker/internal/job"
 
-func Create(name string, p *job.UserParams) (Adapter, error) {
+func Create(name string, p *job.UserParams) (ModelExecutor, error) {
 	panic("not implied")
 }
