+++ internal/job/executer/interface.go
@@ -1,9 +1,10 @@
-package transmit
+package executer
 
 import "github.com/Goboolean/core-system.worker/internal/job"
 
-type Transmitter interface {
+type ModelExecutor interface {
 	job.Common
 
 	SetInput(chan any)
+	Output() chan any
 }
