+++ internal/job/executer/mock-model-exec.go
@@ -1,4 +1,4 @@
-package modelExecute
+package executer
 
 import (
 	"context"
