+++ internal/job/fetcher/factory.go
@@ -1,7 +1,7 @@
-package analyze
+package fetcher
 
 import "github.com/Goboolean/core-system.worker/internal/job"
 
-func Create(name string, p *job.UserParams) (Analyzer, error) {
+func Create(name string, p *job.UserParams) (Fetcher, error) {
 	panic("not implied")
 }
