+++ internal/job/fetcher/fetch-past-stock.go
@@ -1,4 +1,4 @@
-package fetch
+package fetcher
 
 import (
 	"context"
