+++ internal/job/fetcher/fetch-realtime-stock.go
@@ -1,4 +1,4 @@
-package fetch
+package fetcher
 
 import (
 	"context"
