--- internal/job/model-execute/factory.go
@@ -1,7 +1,0 @@
-package modelExecute
-
-import "github.com/Goboolean/core-system.worker/internal/job"
-
-func Create(name string, p *job.UserParams) (ModelExecutor, error) {
-	panic("not implied")
-}
