+++ internal/job/transmitter/factory.go
@@ -1,0 +1,7 @@
+package transmitter
+
+import "github.com/Goboolean/core-system.worker/internal/job"
+
+func Create(name string, p *job.UserParams) (Transmitter, error) {
+	panic("not implied")
+}
